refactor(core): extract hardware_capabilities into proximadb-hardware-caps crate
Completes decomposition contracts step 2's prerequisite: pulls the 1725-LOC rich
hardware-capability layer out of the monolithic root crate into a foundation crate,
after severing its last couplings to `compute`.

GPU-edge inversion (issue #162):
- `detect_gpu()` no longer calls `compute::gpu` directly. The hardware layer holds
  a `GpuDetectorFn` hook (`OnceLock`) populated by `register_gpu_detector`; the
  compute layer registers its detector at startup (src/bin/server.rs) under
  `#[cfg(feature = "gpu")]`, BEFORE `initialize_hardware_capabilities()` — so there
  is no init race. Default / non-gpu builds get empty GPU caps (prior behaviour).
  This removes the `hardware_capabilities -> compute` edge entirely.

Other couplings:
- `HardwareConfig` repointed from `crate::core::config` to `proximadb_config`
  (where it actually lives — `core::config` only re-exported it).

Extraction:
- `hardware_capabilities.rs` (+ its `_tests.rs`) -> `crates/foundation/
  proximadb-hardware-caps`; `src/core/hardware_capabilities.rs` is now a
  `pub use proximadb_hardware_caps::*` shim, so the 53 consumers are unchanged.
- New crate's `gpu` feature mirrors the root's (gates GPU-backend selection +
  Display impls); root `gpu` forwards to it. GPU *detection* is injected, not
  feature-gated in the crate.

CI:
- Pin the macOS `gpu-feature-check` job to `dtolnay/rust-toolchain@1.95.0` (the repo
  pin). `@stable` pulls a newer Rust whose stricter const-eval trips ring 0.17.x's
  aarch64 CAPS_STATIC assertion on macOS (x86_64-linux jobs are unaffected); 1.95.0
  builds gpu cleanly.

Verified on 1.95.0 (aarch64): new crate + tests, default root lib (via shim), and
`--bins --features gpu` (server registration) all green; clippy -D warnings clean.
Unblocks the distance-kernel extraction.
