+++ .github/workflows/ci.yml
@@ -191,7 +191,10 @@ jobs:
     timeout-minutes: 30
     steps:
       - uses: actions/checkout@v4
-      - uses: dtolnay/rust-toolchain@stable
+      # Pin to the repo toolchain (1.95.0): `@stable` lands a newer Rust whose
+      # stricter const-eval trips ring 0.17.x's aarch64 CAPS_STATIC assertion on
+      # macOS (the x86_64-linux jobs are unaffected). 1.95.0 builds gpu cleanly.
+      - uses: dtolnay/rust-toolchain@1.95.0
       - name: Install protobuf
         run: brew install protobuf
       - uses: Swatinem/rust-cache@v2
