+++ Cargo.toml
@@ -31,6 +31,7 @@ members = [
     "crates/foundation/proximadb-encoding-types",
     "crates/foundation/proximadb-filter-expression",
     "crates/foundation/proximadb-hardware",
+    "crates/foundation/proximadb-hardware-caps",
     "crates/foundation/proximadb-kernel",
     "crates/foundation/proximadb-pipeline-operator",
     "crates/foundation/proximadb-proto",
@@ -142,6 +143,7 @@ proximadb-data-model = { path = "crates/foundation/proximadb-data-model" }
 proximadb-config = { path = "crates/foundation/proximadb-config" }
 proximadb-compression-types = { path = "crates/foundation/proximadb-compression-types" }
 proximadb-hardware = { path = "crates/foundation/proximadb-hardware" }
+proximadb-hardware-caps = { path = "crates/foundation/proximadb-hardware-caps" }
 proximadb-kernel = { path = "crates/foundation/proximadb-kernel" }
 proximadb-proto = { path = "crates/foundation/proximadb-proto" }
 proximadb-cache = { path = "crates/foundation/proximadb-cache" }
@@ -268,6 +270,7 @@ proximadb-document-query = { path = "crates/query/proximadb-document-query" }
 proximadb-config = { path = "crates/foundation/proximadb-config" }
 proximadb-compression-types = { path = "crates/foundation/proximadb-compression-types" }
 proximadb-hardware = { path = "crates/foundation/proximadb-hardware" }
+proximadb-hardware-caps = { path = "crates/foundation/proximadb-hardware-caps" }
 proximadb-data-model = { path = "crates/foundation/proximadb-data-model" }
 proximadb-distance-types = { path = "crates/foundation/proximadb-distance-types" }
 proximadb-relational-types = { path = "crates/foundation/proximadb-relational-types" }
@@ -682,7 +685,7 @@ graph-first-sks = []
 datafusion-integration = ["dep:datafusion", "dep:datafusion-common"]
 
 # Hardware acceleration features
-gpu = ["metal"]  # GPU acceleration support (CUDA, ROCm, MPS, OpenCL)
+gpu = ["metal", "proximadb-hardware-caps/gpu"]  # GPU acceleration support (CUDA, ROCm, MPS, OpenCL)
 metal = ["dep:metal"]  # Enable Metal/MPS GPU support on macOS ARM64
 # SIMD optimizations - ARM64 has NEON, not AVX
 simd = []  # Placeholder - would need ARM NEON implementation
